Used the `TestApplication` for unit test cases.
* Extracted `TestApplication` into a separate class so it can be easily reused across unit tests.
