+++ app/src/test/java/org/kiwix/kiwixmobile/localFileTransfer/LocalFileTransferScreenTest.kt
@@ -37,7 +37,7 @@ import org.kiwix.kiwixmobile.core.ui.models.ActionMenuItem
 import org.kiwix.kiwixmobile.core.ui.models.IconItem
 import org.kiwix.kiwixmobile.core.page.SEARCH_ICON_TESTING_TAG
 import org.kiwix.kiwixmobile.core.ui.components.CONTENT_LOADING_PROGRESS_BAR_TESTING_TAG
-import org.kiwix.kiwixmobile.utils.TestApplication
+import org.kiwix.kiwixmobile.TestApplication
 import org.robolectric.RobolectricTestRunner
 import org.robolectric.RuntimeEnvironment
 import org.robolectric.annotation.Config
