+++ app/src/test/java/org/kiwix/kiwixmobile/nav/destination/library/online/OnlineLibraryScreenTest.kt
@@ -52,7 +52,7 @@ import org.kiwix.kiwixmobile.core.ui.models.ActionMenuItem
 import org.kiwix.kiwixmobile.core.utils.ZERO
 import org.kiwix.kiwixmobile.nav.destination.library.online.viewmodel.OnlineLibraryViewModel
 import org.kiwix.kiwixmobile.nav.destination.library.online.viewmodel.OnlineLibraryViewModel.OnlineLibraryUiState
-import org.kiwix.kiwixmobile.utils.TestApplication
+import org.kiwix.kiwixmobile.TestApplication
 import org.kiwix.kiwixmobile.zimManager.libraryView.AvailableSpaceCalculator
 import org.kiwix.kiwixmobile.zimManager.libraryView.LibraryListItem
 import org.robolectric.RobolectricTestRunner
