+++ app/src/test/java/org/kiwix/kiwixmobile/roomDao/HistoryRoomDaoTest.kt
@@ -35,11 +35,12 @@ import org.kiwix.kiwixmobile.KiwixRoomDatabaseTest.Companion.getHistoryItem
 import org.kiwix.kiwixmobile.core.dao.HistoryRoomDao
 import org.kiwix.kiwixmobile.core.data.KiwixRoomDatabase
 import org.kiwix.kiwixmobile.core.page.history.models.HistoryListItem
+import org.kiwix.kiwixmobile.TestApplication
 import org.robolectric.RobolectricTestRunner
 import org.robolectric.annotation.Config
 
 @RunWith(RobolectricTestRunner::class)
-@Config(sdk = [Build.VERSION_CODES.R], application = android.app.Application::class)
+@Config(sdk = [Build.VERSION_CODES.R], application = TestApplication::class)
 class HistoryRoomDaoTest {
   private lateinit var kiwixRoomDatabase: KiwixRoomDatabase
   private lateinit var historyRoomDao: HistoryRoomDao
