+++ app/src/test/java/org/kiwix/kiwixmobile/roomDao/NoteRoomDaoTest.kt
@@ -32,6 +32,7 @@ import org.junit.Test
 import org.junit.Assert.assertEquals
 import org.junit.runner.RunWith
 import org.kiwix.kiwixmobile.KiwixRoomDatabaseTest.Companion.getNoteListItem
+import org.kiwix.kiwixmobile.TestApplication
 import org.kiwix.kiwixmobile.core.dao.NotesRoomDao
 import org.kiwix.kiwixmobile.core.data.KiwixRoomDatabase
 import org.kiwix.kiwixmobile.core.page.notes.models.NoteListItem
@@ -39,7 +40,7 @@ import org.robolectric.RobolectricTestRunner
 import org.robolectric.annotation.Config
 
 @RunWith(RobolectricTestRunner::class)
-@Config(sdk = [Build.VERSION_CODES.R], application = android.app.Application::class)
+@Config(sdk = [Build.VERSION_CODES.R], application = TestApplication::class)
 class NoteRoomDaoTest {
   private lateinit var kiwixRoomDatabase: KiwixRoomDatabase
   private lateinit var notesRoomDao: NotesRoomDao
